requirements.txt
certifi==2020.12.5
dj-database-url==0.5.0
Django==2.2.7
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.7.7
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
